DRY: move check for untested exercises to a reusable funtion
